Refactor main dart and minor section
